requirements/base.txt
Django>=4.1.0<=4.2.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
channels==4.0.0
drf-spectacular==0.26.2
drf_spectacular_sidecar==2023.5.1
Pillow==9.5.0
six==1.16.0
python-decouple==3.8
requests==2.31.0
redis==4.5.5
requests==2.31.0
redis==4.5.5
